+++ src/lib.rs
@@ -91,6 +91,8 @@ pub struct TransformConfig {
     pub loop_limit: u32,
     /// Max length of variable
     pub var_limit: u32,
+    /// Maximum depth of recursion
+    pub depth_limit: u32,
     /// Add source metadata to output
     pub add_metadata: bool,
     /// Default font-size (in user-units)
@@ -114,6 +116,7 @@ impl Default for TransformConfig {
             seed: 0,
             loop_limit: 1000,
             var_limit: 1024,
+            depth_limit: 100,
             add_metadata: false,
             font_size: 3.0,
             font_family: "sans-serif".to_owned(),